Remove stores API
Remove the stores API as the API has now changed to be behind a
search functionality. At this point I see no use case of searching
for stores via this library. Instead, remove the functionality in
its entirety.

In the future we can decide to add back the functionality using the
new search-based API instead.